+++ jgclark.Dashboard/src/react/components/Header/featureFlagItems.js
@@ -16,8 +16,6 @@ const featureFlagSettingDefs = [
   // Note: DBW requests this is kept even when v2.3.0 is released
   { key: 'FFlag_UseTagCache', label: 'Use Tag Cache', description: 'Use Tag Cache to speed up tag/mention searches' },
   { key: 'FFlag_UseTagCacheAPIComparison', label: 'Use Tag Cache API Comparison', description: 'When using Tag Cache, compare the results with the API. (Slows it down.)' },
-  // { key: 'FFlag_IncludeTeamspaceNotes', label: 'Include Teamspace Notes', description: 'Include Teamspace notes in the Dashboard' },
-  // { key: 'FFlag_UseNoteTags', label: 'Use Note Tags', description: 'Use note tags to include whole notes in respective Tag sections' },
   { key: 'FFlag_ShowSectionTimings', label: 'Show Section Timings', description: 'Show timings for how long it took to generate sections' },
 ]