+++ jgclark.Dashboard/src/backupSettings.js
@@ -1,6 +1,6 @@
 // @flow
 //-----------------------------------------------------------------------------
-// Last updated 2025-04-24 for v2.2.2
+// Last updated 2025-10-01 for v2.3.0
 //-----------------------------------------------------------------------------
 
 import moment from 'moment/min/moment-with-locales'
@@ -10,20 +10,18 @@ import { showMessage } from '@helpers/userInput'
 
 //-----------------------------------------------------------------
 
+
 /**
  * Backup Dashboard settings.json file to a dated version in the plugin data folder.
- * @param {boolean} runSilently - If true, don't show any messages.
  */
-export async function backupSettings(runSilently: boolean = false) {
+export async function backupSettings(reason: string = 'backup') {
   try {
     const pluginID = pluginJson['plugin.id']
     const pluginSettings = await DataStore.loadJSON(`../${pluginID}/settings.json`)
-    const backupFilename = `settings_backup_${moment().format('YYYYMMDDHHmmss')}.json`
+    const backupFilename = `settings_${reason}_${moment().format('YYYYMMDDHHmmss')}.json`
     const backupPath = `../${pluginID}/${backupFilename}`
     await DataStore.saveJSON(pluginSettings, backupPath)
-    if (!runSilently) {
-      await showMessage(`Backup of Dashboard settings saved to ${backupPath}`, 'OK', 'Dashboard Settings Backup')
-    }
+    await showMessage(`Backup of Dashboard settings saved to ${backupPath}`, 'OK', 'Dashboard Settings Backup')
     logInfo('backupSettings', `Backup of Dashboard settings saved to ${backupPath}`)
   } catch (error) {
     await showMessage(`Error trying to Backup Dashboard settings. Please see Plugin Console log for details.`, 'OK', 'Dashboard Settings Backup')