+++ jgclark.Dashboard/CHANGELOG.md
@@ -14,6 +14,7 @@ For more details see the [plugin's documentation](https://github.com/NotePlan/pl
 ### Changed
 - When the "filter out lower-priority items" is on, this now calculated across all sections, not just each one independently. The label that says there are hidden items now includes "(click to show all)" text to make it clearer how to turn off the filter for that section.
 - improve cache updates
+- now creates a backup of Dashboard settings whenever installing a new version, whether any settings have changed or not.
 ### Fixed
 - the done count not including items from project notes
 - note-tags when cache is off